Change version in setup.py, separate model file, add description to function
