+++ requirements.txt
@@ -1,3 +1,3 @@
-keras
+keras>=2.0.0
 tensorflow
 pandas
